Move .flattened() to a base class of Event, BadEvent, UnknownBadEvent
